+++ src/main/java/com/github/hcsp/inheritance/Human.java
@@ -1,0 +1,14 @@
+package com.github.hcsp.inheritance;
+
+public class Human extends Animal {
+    private int age;
+
+    public Human(String name, int age) {
+        super(name);
+        this.age = age;
+    }
+
+    public void sayMyAge() {
+        System.out.println("我的年龄是" + this.age);
+    }
+}
